+++ pydiamond/gui/widgets/entry.py
@@ -32,7 +32,7 @@ if TYPE_CHECKING:
     from ...graphics.font import Font
     from ...graphics.renderer import AbstractRenderer
 
-    _TupleFont: TypeAlias = tuple[str | None, int]
+    _TupleFont: TypeAlias = tuple[str | None, float]
     _TextFont: TypeAlias = Font | _TupleFont
